# feat: Using multi_set in shell tool to speedup data copy