--- a/src/shell/config.ini
+++ b/src/shell/config.ini
@@ -49,7 +49,7 @@ worker_priority = THREAD_xPRIORITY_NORMAL
 
 [threadpool.THREAD_POOL_DEFAULT]
 name = default
-worker_count = 8
+worker_count = 20
 
 [threadpool.THREAD_POOL_META_SERVER]
 name = meta_server