### What problem does this PR solve? <!--add issue link with summary if exists-->
https://github.com/apache/incubator-pegasus/issues/923
pegasus_shell copy_data use async_set write now, I want use async_multi_set improve speed sometimes

### What is changed and how does it work?
You can use pegasus_shell exec "copy_data -c cluster_name -a app_name -u"  for it

### The performance tests between async_set and async_multi_set for copy_data:
The same row show different cost from copy the same table to a new empty table. 
|  total_data_count   | row data pre hash_key | async_set (secs) | async_multi_set (secs)  |
|  ----  | ----  |----  | ----  |
| 152,354 | 1   | 8 | 16 |
| 417,408 | 4   | 29 | 11 |
| 657,950 | 50   | 32 | 9 |
| 2,079,400 | 200   | 87 | 18 |
| 1,842,400 | 800   | 399 | 19 |
| 14,374,001 | 2,000   | 631 | 131 |

You should use async_multi_set for copy when your data have multi count per hash_key.