@Arsenal821 2022-03-04
The performance tests between async_set and async_multi_set for copy_data:
The same row show different cost from copy the same table to a new empty table. 
|  case   | seconds for async_set | seconds for async_multi_set  |
|  ----  | ----  | ----  |
| total_data_count: 152354, 1 row data pre hash_key  | 8 | 16 |
| total_data_count: 417408, 4 rows data pre hash_key  | 29 | 11 |
| total_data_count: 657950, 50 rows data pre hash_key  | 32 | 9 |
| total_data_count: 2079400, 200 rows data pre hash_key  | 87 | 18 |
| total_data_count: 1842400, 800 rows data pre hash_key  | 399 | 19 |
| total_data_count: 14374001, 2000 rows data pre hash_key  | 631 | 131 |

You should use async_multi_set for copy when your data have multi count per hash_key.